DATAMONGO-2138 - KDocs; Refactor file structure
