--- spring-data-mongodb/src/test/kotlin/org/springframework/data/mongodb/core/query/TypedQueryExtensionsTest.kt
@@ -1,48 +1,0 @@
-/*
- * Copyright 2010-2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.data.mongodb.core.query
-
-import example.Book
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-/**
- * @author Tjeu Kayim
- */
-class TypedQueryExtensionsTest {
-
-	@Test
-	fun `Call typedQuery with two TypedCriteria`() {
-		val classic = Query(
-			Criteria()
-				.and("price").gt(1100)
-				.and("available").isEqualTo(true)
-		)
-
-		val typed = typedQuery {
-			Book::price gt 1100
-			Book::available isEqualTo true
-		}
-
-		assertEquals(classic.queryObject, typed.queryObject)
-	}
-
-	@Test
-	fun `Typed Query`() {
-		typedQuery { Book::name lte 123 }
-	}
-}
